renamed C++ mem backend
