+++ src/LinAlg/hiopMatrixSparseTriplet.cpp
@@ -13,6 +13,7 @@
 #include <iostream>     // std::cout, std::fixed
 #include <iomanip>      // std::setprecision
 
+#include "MemBackendCppImpl.hpp"
 #include "hiopCppStdUtils.hpp"
 namespace hiop
 {
